hotfix: use correct ABI for fetching super toke balances

+++ src/hooks/useProjectClaimableDonations.ts
@@ -32,7 +32,7 @@ export const useProjectClaimableDonations = (
 				if (balance) {
 					return {
 						token,
-						balance: balance,
+						balance: balance as bigint,
 					};
 				}
 				return null;
